pyproject.toml
[project]
name = "lca-embedding"
version = "0.1.0"
description = "A pipeline for fine-tuning and evaluate embedding model in LCA Domain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.1",
    "deepspeed>=0.18.2",
    "faiss-gpu-cu12>=1.13.0",
    "flagembedding>=1.3.5",
    "gguf>=0.17.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "modelscope>=1.32.0",
    "openai>=2.8.1",
    "peft==0.13.0",
    "python-dotenv>=1.2.1",
    "pytrec-eval>=0.5",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "supabase>=2.25.1",
    "tidas-sdk>=0.1.4",
    "transformers==4.51.3",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple"